+++ src/item.h
@@ -511,7 +511,7 @@ class Item : virtual public Thing
 
 		static std::string getDescription(const ItemType& it, int32_t lookDistance, const Item* item = nullptr, int32_t subType = -1, bool addArticle = true);
 		static std::string getNameDescription(const ItemType& it, const Item* item = nullptr, int32_t subType = -1, bool addArticle = true);
-		static std::string getWeightDescription(const ItemType& it, double weight, uint32_t count = 1);
+		static std::string getWeightDescription(const ItemType& it, uint32_t weight, uint32_t count = 1);
 
 		virtual std::string getDescription(int32_t lookDistance) const;
 		std::string getNameDescription() const;
@@ -556,7 +556,7 @@ class Item : virtual public Thing
 			return items[id].shootRange;
 		}
 
-		virtual double getWeight() const;
+		virtual uint32_t getWeight() const;
 		int32_t getAttack() const {
 			return items[id].attack;
 		}
@@ -760,7 +760,7 @@ class Item : virtual public Thing
 		}
 
 	protected:
-		std::string getWeightDescription(double weight) const;
+		std::string getWeightDescription(uint32_t weight) const;
 
 		Cylinder* parent;
 		ItemAttributes* attributes;
